Added functionality to the core data classes to enable the manager classes to instantiate them more effectively.
git-svn-id: svn+ssh://svn.internal.sanger.ac.uk/repos/svn/arcturus/trunk@719 69345dfe-1762-0410-a7d4-b9dc18918749

+++ arcturus/java/src/uk/ac/sanger/arcturus/data/Core.java
@@ -12,7 +12,7 @@ public class Core {
      * A constant representing an attribute whose value is unknown.
      */
 
-    public final static int UNKNOWN = 0;
+    public final static short UNKNOWN = 0;
 
     private int ID;
     private String name;
